w3c-reports/12092014/mac-wk2/runned-tests/websockets/opening-handshake/001-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:49001/invalid' failed: Invalid status code: BAR

PASS WebSockets: invalid handshake 
